@@ -1,102 +1,99 @@
 {
-    "name": "live-html-previewer",
-    "displayName": "Live HTML Previewer",
-    "description": "Edit and preview HTML documents in VS Code",
-    "version": "0.2.0",
-    "publisher": "hdg",
-    "author": "Harshdeep Gupta",
-    "engines": {
-        "vscode": "^1.0.0"
-    },
-    "categories": [
-        "Other"
-    ],
-    "keywords": [
-        "html",
-        "css",
-        "preview",
-        "vscode"
-    ],
-    "galleryBanner": {
-        "color": "#2E0854",
-        "theme": "dark"
-    },
-    "repository": {
-        "type": "git",
-        "url": "https://github.com/HarshdeepGupta/live-html-preview.git"
-    },
-    "bugs": {
-        "url": "https://github.com/HarshdeepGupta/live-html-preview/issues"
-    },
-    "license": "MIT",
-    "icon": "Resources/Images/icon.png",
-    "activationEvents": [
-        "onLanguage:html"
-    ],
-    "main": "./out/src/extension",
-    "contributes": {
-        "commands": [
-            {
-                "command": "extension.sidePreview",
-                "title": "Show side preview"
-            },
-            {
-                "command": "extension.fullPreview",
-                "title": "Show full preview"
-            },
-            {
-                "command": "extension.inBrowser",
-                "title": "Open in browser"
-            }
-        ],
-        "menus": {
-            "explorer/context": [
-                {
-                    "when": "resourceLangId == html",
-                    "command": "extension.inBrowser",
-                    "group": "navigation"
-                }
-            ],
-            "editor/context": [
-                {
-                    "when": "resourceLangId == html",
-                    "command": "extension.inBrowser",
-                    "group": "navigation"
-                }
-            ]
-        },
-        "keybindings": [
-            {
-                "command": "extension.sidePreview",
-                "key": "ctrl+q s",
-                "mac": "ctrl+q s",
-                "when": "editorTextFocus"
-            },
-            {
-                "command": "extension.fullPreview",
-                "key": "ctrl+q f",
-                "mac": "ctrl+q f",
-                "when": "editorTextFocus"
-            },
-            {
-                "command": "extension.inBrowser",
-                "key": "ctrl+q w",
-                "mac": "ctrl+q w",
-                "when": "editorTextFocus"
-            }
-        ]
-    },
-    "scripts": {
-        "vscode:prepublish": "node ./node_modules/vscode/bin/compile",
-        "compile": "node ./node_modules/vscode/bin/compile -watch -p ./",
-        "postinstall": "node ./node_modules/vscode/bin/install"
-    },
-    "devDependencies": {
-        "typescript": "^1.8.5",
-        "vscode": "^0.11.0"
-    },
-    "dependencies": {
-        "opn": "^4.0.2",
-        "path": "^0.12.7"
-    }
+	"name": "live-tailwind-alpine-preview",
+	"displayName": "Live HTML Previewer with Tailwind/Alpine Support",
+	"description": "Edit and preview HTML documents in VS Code with tailwind and alpinejs support",
+	"version": "1.0.0",
+	"publisher": "n-for-all",
+	"author": "Naji Amer",
+	"engines": {
+		"vscode": "^1.74.0"
+	},
+	"categories": [
+		"Other"
+	],
+	"keywords": [
+		"tailwind",
+		"alpinejs",
+		"html",
+		"css",
+		"preview",
+		"vscode"
+	],
+	"galleryBanner": {
+		"color": "#2E0854",
+		"theme": "dark"
+	},
+	"repository": {
+		"type": "git",
+		"url": "https://github.com/n-for-all/live-tailwind-alpine-preview.git"
+	},
+	"bugs": {
+		"url": "https://github.com/n-for-all/live-tailwind-alpine-preview/issues"
+	},
+	"license": "MIT",
+	"icon": "resources/icon.png",
+	"activationEvents": [
+	],
+	"main": "./out/extension.js",
+	"contributes": {
+		"commands": [
+			{
+				"command": "extension.nj.sidePreview",
+				"title": "Show tailwind side preview"
+			},
+			{
+				"command": "extension.nj.fullPreview",
+				"title": "Show tailwind full preview"
+			}
+		],
+		"menus": {
+			"explorer/context": [
+				{
+					"when": "resourceLangId == html",
+					"command": "extension.nj.fullPreview",
+					"group": "navigation"
+				}
+			],
+			"editor/context": [
+				{
+					"when": "resourceLangId == html",
+					"command": "extension.nj.fullPreview",
+					"group": "navigation"
+				}
+			]
+		},
+		"keybindings": [
+			{
+				"command": "extension.nj.sidePreview",
+				"key": "ctrl+q s",
+				"mac": "ctrl+q s",
+				"when": "editorTextFocus"
+			},
+			{
+				"command": "extension.nj.fullPreview",
+				"key": "ctrl+q f",
+				"mac": "ctrl+q f",
+				"when": "editorTextFocus"
+			}
+		]
+	},
+	"scripts": {
+		"vscode:prepublish": "npm run compile",
+		"compile": "tsc -p ./",
+		"lint": "eslint \"src/**/*.ts\"",
+		"package": "tsc -p ./ && vsce package",
+		"watch": "tsc -watch -p ./"
+	},
+	"dependencies": {
+		"install": "^0.13.0",
+		"path": "^0.12.7"
+	},
+	"devDependencies": {
+		"@types/node": "^18",
+		"@types/vscode": "^1.73.0",
+		"@typescript-eslint/eslint-plugin": "^6.7.0",
+		"@typescript-eslint/parser": "^6.7.0",
+		"eslint": "^8.26.0",
+		"typescript": "^5.4.2"
+	}
 }
