@@ -1,18 +1,14 @@
 "use strict"
 import * as vscode from 'vscode'
 import Utilities from "./Utilities"
-import * as Constants from './Constants'
 
 export default class StatusBarItem {
-
     statusBarItem: vscode.StatusBarItem;
-    utilities: Utilities;
 
-    constructor(utilities?: Utilities) {
+    constructor() {
         this.statusBarItem = vscode.window.createStatusBarItem(vscode.StatusBarAlignment.Left);
-        this.statusBarItem.command = "extension.sidePreview";
-        this.statusBarItem.tooltip = Constants.ExtensionConstants.STATUS_BAR_TOOLTIP;
-        this.utilities = utilities && utilities || new Utilities();
+        this.statusBarItem.command = "extension.nj.sidePreview";
+        this.statusBarItem.tooltip = 'Open Tailwind/Alpine Preview';
     }
 
     updateStatusbar() {
@@ -21,9 +17,8 @@ export default class StatusBarItem {
             this.statusBarItem.hide();
             return;
         }
-        // Only update status if an HTML file
-        if (this.utilities.checkDocumentIsHTML(false)) {
-            this.statusBarItem.text = Constants.ExtensionConstants.STATUS_BAR_TEXT;
+        if (Utilities.isHTML(false)) {
+            this.statusBarItem.text = `$(open-preview)`;
             this.statusBarItem.show();
         }
         else {
